Introduce `Address` newtype for bytecode addresses (#4862)
Replace raw `u32` with a typed `Address` wrapper for all bytecode
address fields in opcodes, handlers, labels, and the compiler. This
provides type safety to distinguish addresses from other u32 values
(register indices, counts, etc.) and displays addresses as hexadecimal
in bytecode dumps.

---------

+++ core/engine/src/vm/mod.rs
@@ -552,7 +552,7 @@ impl Vm {
         let environment_sp = frame.env_fp + handler.environment_count;
 
         // Go to handler location.
-        frame.pc = catch_address;
+        frame.pc = u32::from(catch_address);
 
         self.frame.environments.truncate(environment_sp as usize);
